requirements.txt
python-telegram-bot
requests
firebase_admin